config, main, callbacks, drawing: enable smaller increments of arrows
* change arrowsize & -width to float to enable smaller increments (instead of up to 150 % increase from one size to the next)
* use default value if configured one is too small
* arrowsize = 1 + width = 1 still lead to no arrow at all

+++ src/drawing.c
@@ -41,7 +41,7 @@ void draw_line (GromitData *data,
 void draw_arrow (GromitData *data, 
 		 GdkDevice *dev,
 		 gint x1, gint y1,
-		 gint width,
+		 gfloat width,
 		 gfloat direction)
 {
   GdkRectangle rect;
